# ENH: ndrange, like range, but multidimensional