--- a/doc/release/upcoming_changes/12094.new_function.rst
+++ b/doc/release/upcoming_changes/12094.new_function.rst
@@ -1,0 +1,7 @@
+``ndrange`` a multi-dimensional range-like object
+-------------------------------------------------
+``ndrange`` behaves much like the Python 3 ``range`` object and allows
+multi-dimensional slicing, iteration, reversal and efficient membership lookup.
+``ndrange`` now supersedes the ``ndindex`` object for generating multi-index
+iterators. ``ndindex`` will continue to be maintained for backward
+compatibility.