@hmaarrfk 2018-10-06
I am 200% ok waiting for Jan 1st 2019 for this to be merged if it means I don't need to add Python 2.7 fixes.

But it seems the USE_CHROOT build is also 2.7. Is that correct?
https://travis-ci.org/numpy/numpy/jobs/437922088#L1703