@mattip 2018-10-07
This should go to the mailing list for more general discussion. Please describe there the motivation and a general overview of the way you chose to implement it.