@eric-wieser 2018-10-08
I'm not sure that `ndrange` should have a concept of an order - that sounds like a property of its iterator, just like a `list` object does not have a `reversed` attribute.