@hmaarrfk 2018-10-08
interesting idea,  so the syntax would be:

```
for i in iter(ndrange(arr.shape)[::2], order='F'):
    print(i)
```

I've been debating getting rid of the order parameter all together. I included it because conceptually, some things are easier if you "do them along the first column first".