@hmaarrfk 2018-10-10
Not sure if we need a release note, but this is a draft

```
``ndrange`` a multi-dimensional range-like object
-------------------------------------------------

``ndrange`` now supersedes the ``ndindex`` object for generating multi-index
iterators. ``ndrange`` behaves much like the Python 3 ``range`` object and
allows multi-dimensional slicing, iteration, reversal and efficient containment
lookup. ``ndindex`` will continue to be maintained for backward compatibility.
```