@hmaarrfk 2018-10-10
And here is the patch for fotran ordering. I no longer think it is useful, but maybe somebody else wants a go at it. The tests might be useful.


[ndrange_fortran_ordering_patch_and_tests.patch.txt](https://github.com/numpy/numpy/files/2463082/ndrange_fortran_ordering_patch_and_tests.patch.txt)