@ahaldane 2018-10-10
Look good to me. I would go ahead and add the release note to the changes, and also add a deprecation warning in `np.index`. It's easier for others to evaluate in it's final form.