@hmaarrfk 2018-10-10
@ahaldane thanks for the comments. I'll iterate on the docstring again.

I thought @shoyer mentionned soft deprecating ndindex (i.e. just discouraging the use) in the mailing list.

I don't think it is wise to add a warning in the code to `np.ndindex` since everybody supporting numpy < 1.16 would be unable to use it without catching warnings. Thats super annoying and intrusive.

I think the slicing capabilities of ndrange will be enticing enough for people to switch eventually. Once 1.16 becomes "older", maybe we can add the warning to ndindex?

There has been some discussion about supporting the `__array__` interface. Will have to think of that one.