@hmaarrfk 2018-10-11
I want to leave this here for future developers, my objection to adding `'F'` order indexing is that the implementation I thought of essentially doubles the time it takes to traverse the sequence.

If somebody is so inclined to write the iterator for ndrange in C, then maybe that won't be true anymore and the order parameter of the sequence can be added. 

I don't want to imply in any way that F-order ordering is equivalent in performance to C-order

Finally, if you care about this kind of performance, then this sequence type is probably the wrong thing for you.