@hmaarrfk 2018-11-05
@eric-wieser @ahaldane I think this version addresses many of the concerns you had. It still requires some guidance but it is much closer to a finished product. I'm just sending a friendly ping in case this was still on numpy's radar.