@ahaldane 2018-11-05
Looks good to me modulo the `__array__` and `ravel/flat` discussion which needs a little more discussion.