@hmaarrfk 2018-11-06
@eric-wieser I want to highlight your confusion here because I think it is extremely important.

> Does it represent a sequence of tuples, or an ndarray of tuples? I think that's an important design discussion - are we (using py2 terminology for clarity):
>
>  *  generalizing arange to ndrange as range is to xrange
>  *  Building a completely independent ND generalization of xrange that does not try to mirror ndarray

I 100% agree that `range`, `arange`, `ndrange` are all very similar names that can easily be confused. Users of the library won't know that it is in `index_tricks` and therefore will be even more confused.

Maybe it is best to expand the functionality of `ndindex` since it is pretty unambiguous that it doesn't behave like `np.arange`. 

Rehasing some things that talked about on the mailing list: 2 things stopped me from using `ndindex`
  * input parameter parsing. I think maybe I can just drop the `start` and `step` parameters (which are hard to use anyway) and simply force people to slice the ndindex if they need to.
  * The fact that ndindex itself is the iterator. It should be straightforward to deprecate the `next` method with a warning over a few releases and make the behaviour consistent with `immutable sequences`.

This also has the added benefit of not expanding the numpy namespace.