@hmaarrfk 2018-11-06
finally, the 3rd option is that maybe this function is just too niche, the syntax too confusing, and doesn't belong in such a public facing API.