@hmaarrfk 2019-02-01
@ahaldane @eric-wieser is there still any interest in this?

Is it better to refactor `ndindex` instead?