@hmaarrfk 2019-05-19
Did we want to benchmark this and simply merge in the performance enhancements into ndindex?