@hmaarrfk 2024-07-03
You know I was just thinking about this. 

I think now that 2.7 is gone I can take a stab at it.