@mattip 2024-09-04
Did this ever reach the mailing list as an API change?