@hmaarrfk 2024-09-04
I would love to run my benchmark again in 2024, but I keep running into:

```
Using pip 24.2 from /home/mark/miniforge3/envs/np/lib/python3.12/site-packages/pip (python 3.12)
Non-user install because site-packages writeable
Created temporary directory: /tmp/pip-build-tracker-5gzlhngk
Initialized build tracking at /tmp/pip-build-tracker-5gzlhngk
Created build tracker: /tmp/pip-build-tracker-5gzlhngk
Entered build tracker: /tmp/pip-build-tracker-5gzlhngk
Created temporary directory: /tmp/pip-install-6hyb4gpy
Created temporary directory: /tmp/pip-ephem-wheel-cache-579d6re2
Obtaining file:///home/mark/git/numpy
  Added file:///home/mark/git/numpy to build tracker '/tmp/pip-build-tracker-5gzlhngk'
  Running command Checking if build backend supports build_editable
  Checking if build backend supports build_editable ... done
  Created temporary directory: /tmp/pip-modern-metadata-z_g4s_0q
  Running command Preparing editable metadata (pyproject.toml)

  meson-python: error: Could not find meson version 0.63.3 or newer, found .
  error: subprocess-exited-with-error
```

I installed my environment with:

```
# using conda-forge.
mamba create --name np python=3.12 meson-python ninja pkg-config python-build cytho
n libblas libcblas liblapack  compilers
```

tips appreciated