@hmaarrfk 2024-09-04
I used the power of copy and paste to obtain the following local benchmarks:

```python
In [11]: In [3]: %%timeit
    ...:    ...: for it in np.ndrange((100, 100)):
    ...:    ...:     pass
    ...:
143 μs ± 2.51 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [12]:
    ...: In [4]: %%timeit
    ...:    ...: for it in np.ndindex((100, 100)):
    ...:    ...:     pass
    ...:    ...:

1.15 ms ± 16.4 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [13]:

In [13]: np.__version__
'2.1.1'

In [14]: import sys

In [15]: sys.version
'3.10.14 | packaged by conda-forge | (main, Mar 20 2024, 12:45:18) [GCC 12.3.0]'
```