@mattip 2024-09-05
@ahaldane @shoyer @eric-wieser as people involved in the original mailing list thread, any thoughts?