@shoyer 2024-09-05
In the many years since that mailing list discussion, I've gotten a bit more conservative in my taste for API design.

It is still not clear to me who would use this feature. It feels like a solution that exists for the sake of "elegance" or "symmetry," not actual user needs.

If the intended users are library authors doing indexing manipulations for arrays,  then @asmeurer's [ndindex](https://github.com/Quansight-Labs/ndindex) is a much more complete solution.